Delete, Edit Store

+++ backend/app/controllers/store_controller.ts
@@ -1,5 +1,5 @@
 import type { HttpContext } from '@adonisjs/core/http'
-import { createStoreValidator } from '#validators/store'
+import { createStoreValidator, editStoreValidator } from '#validators/store'
 import Store from '#models/store'
 import User from '#models/user'
 
@@ -72,17 +72,26 @@ export default class StoreController {
   }
 
   // Delete store by ID
-  async delete({ params, response }: HttpContext) {
+  async delete({ params, response, auth }: HttpContext) {
     const { id } = params
     try {
+      await auth.check()
+      const user = auth.user
       const store = await Store.find(id)
       if (!store) {
         return response.status(404).json({
           error: 'Store not found',
         })
       }
+      if (!user) {
+        return response.status(404).json({
+          error: 'User not found',
+        })
+      }
       await store.delete() // Elimina la tienda
-      return response.status(204).send('') // Retorna un 204 No Content
+      user.isSeller = 'false' // Cambia el estado del usuario a no vendedor
+      await user.save() // Guarda el usuario
+      return response.status(204) // Retorna un 204 No Content
     } catch (error) {
       return response.status(500).json({
         message: 'Error deleting store',
@@ -90,5 +99,49 @@ export default class StoreController {
       })
     }
   }
-}
 
+  async edit({ params, request, response, auth }: HttpContext) {
+    const { id } = params
+    try {
+      const { name, description, portraitImage, storeImage } =
+        await request.validateUsing(editStoreValidator)
+      const store = await Store.find(id)
+      await auth.check()
+      const user = auth.user
+
+      if (!store) {
+        return response.status(404).json({
+          error: 'Store not found',
+        })
+      }
+      if (!user) {
+        return response.status(404).json({
+          error: 'User not found',
+        })
+      }
+      if (store.ownerId !== user.id) {
+        return response.status(403).json({
+          error: 'Unauthorized',
+        })
+      }
+
+      if (name) {
+        store.name = name
+      }
+      if (description) {
+        store.description = description
+      }
+      if (portraitImage) {
+        store.portraitImage = portraitImage
+      }
+      if (storeImage) {
+        store.storeImage = storeImage
+      }
+
+      await store.save()
+      return response.status(200).json(store)
+    } catch (e) {
+      console.log(e)
+    }
+  }
+}
